Organized file structure to have UI and Loadout packages Fixed imports and references which changed due to the move
