+++ src/test/kotlin/AsteroidDemo.kt
@@ -6,6 +6,7 @@ import com.simsilica.mathd.Vec3d
 import com.simsilica.sim.AbstractGameSystem
 import com.simsilica.sim.SimTime
 import universe.*
+import universe.ui.*
 
 /*
  * A simple demo for interacting with a randomly generated asteroid field
