+++ src/test/kotlin/LoadoutEditorDemo.kt
@@ -1,5 +1,9 @@
 import com.jme3.system.AppSettings
 import universe.*
+import universe.loadout.CargoEquip
+import universe.loadout.EngineEquip
+import universe.loadout.cacheEquipment
+import universe.ui.LoadoutEditorState
 
 class LoadoutEditorDemo: SpaceTraderApp(false){
     override fun simpleInitApp() {
